requirements.txt
flask
line-bot-sdk
google-generativeai
firebase-admin
gunicorn
python-dotenv
